feat: implement remainings
