+++ src/components/features/admin/Forecasts/ForecastForm/RecentAvalanchesSection/RecentAvalanchesSection.tsx
@@ -1,76 +1,34 @@
-import { useCallback, useState } from 'react'
-import { useBoolean } from '@components/hooks'
 import { Icon } from '@components/icons'
 import { Button } from '@components/ui'
-import type { Avalanche } from '@domain/types'
+import type { RegionId } from '@domain/types'
 import { useTranslations } from 'next-intl'
-import { useFieldArray, useFormContext } from 'react-hook-form'
 
 import { AvalancheForm } from './AvalancheForm'
 import { AvalanchePickerModal } from './AvalanchePickerModal'
 import { AvalanchesList } from './AvalanchesList'
-import type { AvalancheFormState } from '../common'
+import useRecentAvalanchesSection from './useRecentAvalanchesSection'
 import { initialAvalancheData } from '../constants'
-import type { ForecastFormSchema } from '../schema'
 
-const toDate = (value: Date | string | null): Date | null =>
-  !value ? null : value instanceof Date ? value : new Date(value)
+type RecentAvalanchesSectionProps = {
+  forecastId?: number
+  regionId: RegionId
+}
 
-const RecentAvalanchesSection = ({ forecastId }: { forecastId?: number }) => {
+const RecentAvalanchesSection = ({ forecastId, regionId }: RecentAvalanchesSectionProps) => {
   const t = useTranslations()
-  const { control } = useFormContext<ForecastFormSchema>()
-  const { append, fields, remove, update } = useFieldArray({
-    control,
-    keyName: 'localId',
-    name: 'recentAvalanches',
-  })
-
-  const [formState, setFormState] = useState<AvalancheFormState>(null)
-  const [isPickerOpen, { setFalse: closePicker, setTrue: openPicker }] = useBoolean(false)
-
-  const handleCreateFormOpen = useCallback(() => setFormState({ mode: 'create' }), [])
-  const handleFormClose = useCallback(() => setFormState(null), [])
-
-  const handleSubmit = useCallback(
-    (data: Omit<Avalanche, 'regionId'>) => {
-      const prepared = { ...data, date: toDate(data.date) }
-
-      if (formState?.mode === 'edit' && formState.localId) {
-        const index = fields.findIndex((field) => field.localId === formState.localId)
-
-        if (index === -1) return
-        update(index, prepared)
-      } else {
-        append(prepared)
-      }
-
-      handleFormClose()
-    },
-    [append, fields, formState, handleFormClose, update],
-  )
-
-  const handleDelete = useCallback(
-    (localId: string) => {
-      const index = fields.findIndex((field) => field.localId === localId)
-
-      if (index === -1) return
-      remove(index)
-    },
-    [fields, remove],
-  )
-
-  const handlePickerConfirm = useCallback(
-    (selected: Avalanche[]) => {
-      const existingIds = new Set(fields.map((field) => field.id).filter((id) => id != null))
-
-      selected
-        .filter((avalanche) => !avalanche.id || !existingIds.has(avalanche.id))
-        .forEach((avalanche) => append({ ...avalanche, date: toDate(avalanche.date) }))
-
-      closePicker()
-    },
-    [append, closePicker, fields],
-  )
+  const {
+    closePicker,
+    fields,
+    formState,
+    handleCreateFormOpen,
+    handleDelete,
+    handleFormClose,
+    handlePickerConfirm,
+    handleSubmit,
+    isPickerOpen,
+    openPicker,
+    setFormState,
+  } = useRecentAvalanchesSection()
 
   return (
     <section className="flex flex-col gap-4">
@@ -110,6 +68,7 @@ const RecentAvalanchesSection = ({ forecastId }: { forecastId?: number }) => {
         isOpen={isPickerOpen}
         onClose={closePicker}
         onConfirm={handlePickerConfirm}
+        regionId={regionId}
       />
     </section>
   )
